cleanup some tests whose coverage has been duplicated
